+++ src/main/java/faang/school/projectservice/service/StageInvitationService.java
@@ -3,8 +3,7 @@ package faang.school.projectservice.service;
 import faang.school.projectservice.dto.stage_invitation.StageInvitationDto;
 import faang.school.projectservice.dto.stage_invitation.StageInvitationFilterDto;
 import faang.school.projectservice.filter.stage_invitation.StageInvitationFilter;
-import faang.school.projectservice.mapper.stage_invitation.StageInvitationMapper;
-import faang.school.projectservice.model.TeamMember;
+import faang.school.projectservice.mapper.StageInvitationMapper;
 import faang.school.projectservice.model.stage_invitation.StageInvitation;
 import faang.school.projectservice.model.stage_invitation.StageInvitationStatus;
 import faang.school.projectservice.repository.StageInvitationRepository;
